further refactoring of per-component imports to reflect new components directory structure
