+++ src/components/home-view/home-view.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import GeneratorForm from '../generator-form/generator-form';
 // import LibraryCarousel from './library-carousel';
 
-import { fetchClouds } from '../actions';
+import { fetchClouds } from '../../actions';
 
 export function HomeView(props) {
   return (
